Finished market info and contract history pages and their controllers. Small modifications to existing pages.
Todo: Make a useful index page that allows searching for Markets/Contracts containing search term(s), other useful features.

Todo: Style pages.

Todo: Ideally someday the REST api would be managed by an app and the web frontend would be dynamic, probably from Angular, and the building/managing of the DOM would be handled by the user's browser, not built by Thymeleaf on the server side, but this is just for now.